# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is docs of Astria Explorer API.",
        "title": "Swagger Astria Explorer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.astrium.io",
    "paths": {
...
        "/v1/stats/rollup/series/{hash}/{name}/{timeframe}": {
            "get": {
                "description": "Get histogram with precomputed rollup by series name and timeframe",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get histogram with precomputed rollup stats",
                "operationId": "stats-rollup-series",
                "parameters": [
                    {
                        "maxLength": 48,
                        "minLength": 48,
                        "type": "string",
                        "description": "Hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "hour",
                            "day",
                            "month"
                        ],
                        "type": "string",
                        "description": "Timeframe",
                        "name": "timeframe",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "size",
                            "avg_size",
                            "min_size",
                            "max_size",
                            "actions_count"
                        ],
                        "type": "string",
                        "description": "Series name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Time from in unix timestamp",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Time to in unix timestamp",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.RollupSeriesItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/v1/stats/series/{name}/{timeframe}": {
            "get": {
                "description": "Get histogram with precomputed stats by series name and timeframe",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get histogram with precomputed stats",
                "operationId": "stats-series",
                "parameters": [
                    {
                        "enum": [
                            "hour",
                            "day",
                            "month"
                        ],
                        "type": "string",
                        "description": "Timeframe",
                        "name": "timeframe",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "data_size",
                            "tps",
                            "bps",
                            "rbps",
                            "fee",
                            "supply_change",
                            "block_time",
                            "tx_count",
                            "bytes_in_block",
                            "gas_price",
                            "gas_efficiency",
                            "gas_used",
                            "gas_wanted"
                        ],
                        "type": "string",
                        "description": "Series name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Time from in unix timestamp",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Time to in unix timestamp",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.SeriesItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/v1/stats/summary": {
            "get": {
                "description": "Get network summary",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get network summary",
                "operationId": "stats-summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.NetworkSummary"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "responses.NetworkSummary": {
            "type": "object",
            "properties": {
                "block_time": {
                    "type": "number",
                    "format": "number",
                    "example": 2345
                },
                "bps": {
                    "type": "number",
                    "format": "number",
                    "example": 0.17632
                },
                "bytes_in_block": {
                    "type": "integer",
                    "format": "integer",
                    "example": 1024
                },
                "data_size": {
                    "type": "integer",
                    "format": "integer",
                    "example": 1000000
                },
                "fee": {
                    "type": "string",
                    "format": "string",
                    "example": "1012012"
                },
                "rbps": {
                    "type": "number",
                    "format": "number",
                    "example": 0.17632
                },
                "supply": {
                    "type": "string",
                    "format": "string",
                    "example": "1029129"
                },
                "tps": {
                    "type": "number",
                    "format": "number",
                    "example": 0.17632
                },
                "tx_count": {
                    "type": "integer",
                    "format": "integer",
                    "example": 100
                }
            }
        },
...
        "responses.RollupSeriesItem": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "value": {
                    "type": "string",
                    "format": "string",
                    "example": "0.17632"
                }
            }
        },
...
        "responses.SeriesItem": {
            "type": "object",
            "properties": {
                "max": {
                    "type": "string",
                    "format": "string",
                    "example": "0.17632"
                },
                "min": {
                    "type": "string",
                    "format": "string",
                    "example": "0.17632"
                },
                "time": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "value": {
                    "type": "string",
                    "format": "string",
                    "example": "0.17632"
                }
            }
        },
...
    }
}